Start a new module Properties, but keep the old version in Properties-old.
This commit breaks brsim :-)

I will be reimplementing Properties from scratch in the next series of
commits.  I would like to use FGL, which should be a lot easier than
Data.Graph, especially in what concerns building the subgraphs.  However,
the functions which do not depend on the graph library in use should stay
the same, that is why I am not ditching the old code altogether.